Start writing table/parameterized test
Previously, individual tests were written which was what I was used to.
However I believe the assignment required a different style of test,
namely table tests. I researched and found the alternative parameterized
test style for Playwright. This test takes an array of inputs and
repeats the same test for each input, with branches such that different
scenarios are handled.